thumbnail-service/app/requirements.txt
flask
pika
boto3
Pillow
requests
